sites/echo/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Background Tasks
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
httpx==0.26.0
aiohttp==3.9.1
feedparser==6.0.10
newspaper3k==0.2.8
selenium==4.17.2
webdriver-manager==4.0.1

# Stock Market Data
yfinance==0.2.36
alpha-vantage==2.3.1
pandas-ta==0.3.14b0
finnhub-python==2.4.19

# LLM APIs
google-generativeai==0.3.2
openai==1.9.0
anthropic==0.8.1
groq==0.4.1
together==0.2.10

# Data Processing & Analysis
pandas==2.1.4
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0

# Technical Analysis
ta==0.11.0
talib-binary==0.4.26; sys_platform != 'darwin'

# NLP & Sentiment
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.37.0
torch==2.1.2

# Statistics & Math
statsmodels==0.14.1
arch==6.2.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
tenacity==8.2.3
loguru==0.7.2
rich==13.7.0
jinja2==3.1.3
orjson==3.9.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Date/Time
python-dateutil==2.8.2
pytz==2024.1
